% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsimg_gif.R
\name{displayGif}
\alias{displayGif}
\title{Display a GIF using system default viewer}
\usage{
displayGif(gif_data, viewer = "default", cleanup = TRUE)
}
\arguments{
\item{gif_data}{Raw vector of GIF data or path to a GIF file}

\item{viewer}{Character. Display method: "default", "browser", "tempfile". Default: "default"}

\item{cleanup}{Logical. Remove temporary file after viewing. Default: TRUE}
}
\value{
Invisible list with file path and content type
}
\description{
Display a GIF using system default viewer
}
\examples{
data(volcano)
plots <- lapply(c(0, 45, 90, 135), function(angle) {
  persp(volcano, theta = angle, phi = 30,
        col = "lightblue", shade = 0.5)
})
gif_data <- createGif(plot_df = data.frame(plot = I(plots)))

\dontrun{
  displayGif(gif_data)
  
  displayGif(gif_data, viewer = "browser")
  
  saveGif(gif_data, "volcano.gif")
  displayGif("volcano.gif")
}
}
