% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsimg_gif.R
\name{saveGif}
\alias{saveGif}
\title{Save GIF data to file}
\usage{
saveGif(gif_data, file_path)
}
\arguments{
\item{gif_data}{Raw vector of GIF data}

\item{file_path}{Output file path}
}
\value{
Invisible NULL
}
\description{
Save GIF data to file
}
\examples{
# gif_data <- createGif("image.png")
# saveGif(gif_data, "output.gif")
}
