% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsimg_gif.R
\name{createGif}
\alias{createGif}
\title{Create a GIF from a sequence of images or plots}
\usage{
createGif(
  image_paths = NULL,
  debug = FALSE,
  plot_df = NULL,
  width = 800,
  height = 600
)
}
\arguments{
\item{image_paths}{Optional. A character vector containing the file paths of the images to include in the GIF. If not provided, `plot_df` must be supplied.}

\item{debug}{Logical. If TRUE, prints debug information during the GIF creation process. Default is FALSE.}

\item{plot_df}{Optional. A dataframe containing plot objects to be included in the GIF. Each row should contain a plot object. If not provided, `image_paths` must be supplied.}

\item{width}{Optional. The width of the output image in pixels if `plot_df` is provided. Default is 800.}

\item{height}{Optional. The height of the output image in pixels if `plot_df` is provided. Default is 600.}
}
\value{
A raw vector representing the GIF data.
}
\description{
This function generates a GIF from a sequence of image files or plot objects.
}
\examples{
# Create a GIF from a sequence of image files
createGif(c("image1.png", "image2.png"))

# Create a GIF from a dataframe of plots
library(ggplot2)
plot_df <- data.frame(plot = I(list(
  ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point(),
  ggplot(mtcars, aes(x = hp, y = mpg)) + geom_point()
)))
createGif(plot_df = plot_df)
}
