% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsimg_gif.R
\name{saveGif}
\alias{saveGif}
\title{Save GIF data to file}
\usage{
saveGif(gif_data, file_path)
}
\arguments{
\item{gif_data}{Raw vector of GIF data}

\item{file_path}{Output file path}
}
\value{
Invisible NULL
}
\description{
Save GIF data to file
}
\examples{
# Criar um GIF de exemplo com plots
data(volcano)
plots <- lapply(c(30, 60, 90), function(angle) {
  persp(volcano, theta = angle, phi = 30,
        col = "lightblue", shade = 0.5)
})
gif_data <- createGif(plot_df = data.frame(plot = I(plots)))

\dontrun{
  # Salvar o GIF em arquivo
  saveGif(gif_data, "volcano_rotation.gif")
}
}
