% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsimg_gif.R
\name{saveGif}
\alias{saveGif}
\title{Save the raw vector into a proper .gif file}
\usage{
saveGif(gif_data, file_path)
}
\arguments{
\item{gif_data}{A raw vector representing the GIF data.}

\item{file_path}{The file path where the GIF should be saved with its name}
}
\description{
Save the raw vector into a proper .gif file
}
\examples{
# saveGif(gif_data, "output.gif")
}
